Auto-detect and allow pass-by-value FT_Vector to FT_Glyph_To_Bitmap
The 3rd argument to FT_Glyph_To_Bitmap should be a pointer to FT_Vector
(i.e. pass-by-reference), but a pass-by-value FT_Vector is more
python-natural, and the examples (examples/glyph-color.py,
examples/glyph-outline.py) have always been wrong...
So we auto-detect the wrong calling and make that case work anyway.

This was discussed in:
https://github.com/rougier/freetype-py/issues/51