tests: extend the PyInstaller hook test
Have the PyInstaller hook test ensure that the freetype shared
library is bundled with the frozen application. Otherwise, the
frozen application might still work in spite of defunct hook due
to the fact that the system-wide freetype shared library is
available (which is likely true in the case of the CI system).